requirements/docker_local.txt
django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.4
pyjwt
psycopg2==2.9.5